requirements.txt
beautifulsoup4==4.13.4
duckduckgo_search==8.0.1
google-ai-generativelanguage==0.6.15
google-genai==1.13.0
google-generativeai==0.8.5
llama-cloud==0.1.19
llama-cloud-services==0.6.21
llama-index==0.12.34
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.34.post1
llama-index-embeddings-google-genai==0.1.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.11
llama-index-llms-google-genai==0.1.10
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.21
nltk==3.8.1
numpy==2.2.5
streamlit
openai==1.77.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==10.4.0
pypdf==5.4.0
pypdfium2==4.30.1
python-dotenv==1.1.0
python-pptx==1.0.2
PyYAML==6.0.2
requests==2.32.3
tqdm==4.67.1

